move makeResponse to state-machine library
