+++ apps/newsletters-api/src/app/state-machine.ts
@@ -1,12 +1,9 @@
 import { newslettersWorkflowStepLayout } from '@newsletters-nx/newsletter-workflow';
 import type {
 	CurrentStepRouteRequest,
-	CurrentStepRouteResponse,
 	WizardStepData,
-	WizardStepLayout,
 } from '@newsletters-nx/state-machine';
 import {
-	convertWizardStepLayoutButtonsToWizardButtons,
 	setupInitialState,
 	stateMachineButtonPressed,
 } from '@newsletters-nx/state-machine';
@@ -27,27 +24,3 @@ export async function getState(
 		  )
 		: await setupInitialState(requestBody, storageInstance);
 }
-
-export const getResponseFromBodyAndStateAndWizardStepLayout = (
-	requestBody: CurrentStepRouteRequest,
-	state: WizardStepData,
-	nextWizardStepLayout: WizardStepLayout,
-): CurrentStepRouteResponse => {
-	const { staticMarkdown, dynamicMarkdown } = nextWizardStepLayout;
-
-	const markdown = dynamicMarkdown
-		? dynamicMarkdown(requestBody.formData, state.formData)
-		: staticMarkdown;
-
-	const currentStepRouteResponse = {
-		markdownToDisplay: markdown,
-		currentStepId: state.currentStepId,
-		buttons: convertWizardStepLayoutButtonsToWizardButtons(
-			nextWizardStepLayout.buttons,
-		),
-		errorMessage: state.errorMessage,
-		formData: state.formData,
-	};
-
-	return currentStepRouteResponse;
-};
